+++ app/command/deploy.go
@@ -1,7 +1,7 @@
 package command
 
 import (
-	"github.com/go-xiaohei/pugo/ext/deploy"
+	"github.com/go-xiaohei/pugo/app/extend/deploy"
 	"github.com/urfave/cli"
 )
