+++ app/extend/deploy/qiniu.go
